feat: Refactor nominee handling by creating NomineeController and NomineeService
